# swagger/openapi.yaml
openapi: 3.0.3
info:
  title: FMS API
  version: 1.0.0
  description: |
    Fleet Management System API - versi awal.
    Fitur utama:
    - Manajemen kendaraan
    - Posisi terkini kendaraan
    - Trip history & alert (read-only dulu)
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/users:
    post:
      summary: Create user in organization (ORG_ADMIN)
      tags: [Users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
    get:
      summary: List users
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: organizationId
          schema:
            type: integer
          required: false
      responses:
        '200':
          description: List users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserDTO'
  /api/users/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
    get:
      summary: Get user by id
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
    put:
      summary: Update user
      tags: [Users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
    delete:
      summary: Soft-delete user
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
...
components:
  schemas:
...
    CreateUserRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        role:
          type: string
    UpdateUserRequest:
      type: object
      properties:
        fullName:
          type: string
        password:
          type: string
        role:
          type: string
        active:
          type: boolean
    UserDTO:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        userType:
          type: string
        organizationId:
          type: integer
        orgRole:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT